Refactor: Extract LiveKit room connector logic into hooks
This commit refactors the `LivekitRoomConnector` component by extracting its core logic into dedicated hooks: `useAgentDispatch`, `useRoomEvents`, and `useLivekitConnection`. This improves code organization, testability, and reusability.

Additionally, a new `RoomConnectorUI` component is introduced to handle the UI presentation, further separating concerns. The documentation in `docs/code-organization.md` is updated to reflect these changes and the completed refactoring of the LiveKit connector.
